Fix OWNERS.
adelva doesn't use/need this, so he doesn't own it. bohu (and the
emulator team) only care about the host, so anything that affects them
is NSI anyway. So it seems like the "interesting" user here from a
security perspective is vold, and OWNERS should probably reflect that
(rather than just being what a script once said were the two most recent
humans to have touched a file in here).

Bug: http://b/271932190
Test: treehugger
Change-Id: Ib872a324b68809b2c7404b3a02193b261aa9ad01