Build sgdisk with _FILE_OFFSET_BITS=64.
Since
https://sourceforge.net/p/gptfdisk/code/ci/7dfa8984f5a30f313d8675ff6097c8592d636d10/
upstream assumes that it's being built with _FILE_OFFSET_BITS=64.
This is true for 64-bit Android, but not for 32-bit Android. We can't
change that globally because it would be a source-incompatible ABI
change (_all_ code that uses off_t or an associated function would have
to be recompiled with _FILE_OFFSET_BITS=64, including code not in our
tree).

We can safely make this change here, though, because the "library" is
actually just to reduce duplication between the sgdisk binary and the
fuzzer. We could probably get rid of the library completely and just
rely on cc_defaults or a filegroup for this, but the less invasive
change is just to add visibility clauses (and a comment!) to ensure that
no-one starts exporting anything from here (even though the GPL license
makes that unlikely anyway).

While I'm here, we can remove the Darwin special case (which was a bit
weird, tbh, given that Darwin has _always and only_ had a 64-bit off_t).

Bug: https://issuetracker.google.com/298312533 (cherrypick request)
Bug: https://issuetracker.google.com/294327337
Test: mm for both a 32-bit lunch and a 64-bit lunch
(cherry picked from https://android-review.googlesource.com/q/commit:f123bad00c455f602e7e26e2d6f5ec58ab2754a7)
Merged-In: Icc740cbcf2a0e8747876d1e725f5439c5795a9b2
Change-Id: Icc740cbcf2a0e8747876d1e725f5439c5795a9b2